// frontend/src/model/types.ts
export enum UserRole {
    Customer = 'Customer',
    Pilot = 'Pilot',
    Admin = 'Admin'
}

export enum ShipmentStatus {
    WaitingForApproval = 'WaitingForApproval',
    Approved = 'Approved',
    Denied = 'Denied',
    Assigned = 'Assigned',
    InTransit = 'InTransit',
    Delivered = 'Delivered',
    Cancelled = 'Cancelled'
}

export enum RiskLevel {
    Low = 'Low',
    Medium = 'Medium',
    High = 'High',
    Critical = 'Critical'
}